cleaned up code and added a custom hook component for usage in App.js. Added config file and folder which stores API_URL instead of having it repeat in every component file.
