+++ frontend/src/components/AddGrocery.js
@@ -1,7 +1,6 @@
 import { useState } from "react";
+import { API_URL } from "../config/config.js"; 
 import axios from "axios";
-
-const API_URL = "https://render-wq7v.onrender.com/groceries";
 
 function AddGrocery({ addGrocery }) {
   const [newGrocery, setNewGrocery] = useState({ name: "", price_per_pound: "", units_available: "" });
