+++ frontend/src/components/DeleteGrocery.js
@@ -1,6 +1,5 @@
+import { API_URL } from "../config/config.js"; 
 import axios from "axios";
-
-const API_URL = "https://render-wq7v.onrender.com/groceries";
 
 function DeleteGrocery({ id, onDelete }) {
   const handleDelete = () => {
