+++ frontend/src/config/config.js
@@ -1,0 +1,1 @@
+export const API_URL = "https://render-wq7v.onrender.com/groceries";
